backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

sqlmodel==0.0.22
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

pydantic==2.10.3
pydantic-settings==2.7.0

kafka-python==2.0.2

mlflow==2.19.0
scikit-learn==1.5.2
numpy==2.1.3


python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
